updated the playing screen as well as the delay algorithm

+++ BandController/lib/Song/Song.cpp
@@ -30,7 +30,7 @@ Song::Song()
 // Setter
 void      Song::setPath(String path) {_path = path;}
 void      Song::setTitle(String title) {_title = title;}
-void      Song::setLength(uint16_t length) {_length = length;}
+void      Song::setLength(uint32_t length) {_length = length;}
 
 void      Song::setFormat(uint16_t format) {_format = format;}
 void      Song::setNtrks(uint16_t ntrks) {_ntrks = ntrks;}
@@ -40,7 +40,7 @@ void      Song::setUsPerMidiQn(uint32_t usPerMidiQn) {_usPerMidiQn = usPerMidiQn
 // Getter
 String    Song::getPath() {return _path;}
 String    Song::getTitle() {return _title;}
-uint16_t  Song::getLength() {return _length;}
+uint32_t  Song::getLength() {return _length;}
 
 uint16_t  Song::getFormat() {return _format;}
 uint16_t  Song::getNtrks(){return _ntrks;}
